% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userfunctions.R
\name{alignAllReads}
\alias{alignAllReads}
\title{Align All Reads}
\usage{
alignAllReads(bamf, gnm)
}
\arguments{
\item{bamf}{path to BAM or SAM file}

\item{gnm}{which reference genome to use, accepted options are "hg19",
"hg38", or "GRCh38"}
}
\value{
Data frame containing the aligned reads where each row is
a read alignment and the three columns are CIGAR, Ref., and Read.
}
\description{
This function aligns all reads from a BAM or SAM file that don't
contain NA to the reference genome and returns the resulting
alignment in a DataFrame. In the case of a SAM file, the function
will first create a corresponding BAM file, which will be saved in
the same location as the SAM file.
}
\examples{
alignAllReads(system.file("extdata", "ex_bam.bam", 
                        package = "readAlign"),"hg38")
}
\seealso{
\code{\link{alignRead}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
