% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{alignment}
\alias{alignment}
\title{Alignment}
\usage{
alignment(cigar, cigardf, read, ref, makedf)
}
\arguments{
\item{cigar}{CIGAR string associated with a read}

\item{cigardf}{dataframe of CIGAR string contents from cigarreader}

\item{read}{read sequence from BAM file}

\item{ref}{reference sequence from genome corresponding to read}

\item{makedf}{bool indicating if a dataframe of the results should be made}
}
\value{
Nothing
}
\description{
This function sends the read and reference sequence to be aligned
based on the CIGAR dataframe and either prints the alignment of
a single read or stores the alignments of the reads
}
\seealso{
\code{\link{cigarReader}}\cr
\code{\link{alterReads}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
