% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userfunctions.R
\name{alignRead}
\alias{alignRead}
\title{Align Read}
\usage{
alignRead(bamf, gnm, readidx)
}
\arguments{
\item{bamf}{path to BAM or SAM file}

\item{gnm}{which reference genome to use, accepted options are "hg19",
"hg38", or "GRCh38"}

\item{readidx}{index number of read to align}
}
\value{
DataFrame containing the results of the alignment
}
\description{
This function aligns one specified read from a BAM or SAM file to
the reference genome and stores the resulting alignment in a
DataFrame along with printing it. Function will return an error in
the case that the CIGAR string is not available for the read. In
the case of a SAM file, the function will first create a
corresponding BAM file, which will be saved in the same location
as the SAM file.
}
\examples{
alignRead(system.file("extdata", "ex_bam.bam", 
                        package = "readAlign"),"hg38", 1)
}
\seealso{
\code{\link{alignAllReads}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
