% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{alterReads}
\alias{alterReads}
\title{Read Alter-er}
\usage{
alterReads(idx, cdf)
}
\arguments{
\item{idx}{relevant row in CIGAR dataframe}

\item{cdf}{CIGAR dataframe, output from cigarreader}
}
\value{
Nothing
}
\description{
This function takes the CIGAR dataframe and a row in the
dataframe that indicates changes needed for alignment, and alters
the read and reference sequences as necessary based on the CIGAR.
}
\seealso{
\code{\link{cigarReader}}\cr
\code{\link{alignment}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
