% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{cigarReader}
\alias{cigarReader}
\title{CIGAR DataFrame}
\usage{
cigarReader(cigar)
}
\arguments{
\item{cigar}{CIGAR string associated with a read}
}
\value{
DataFrame where each row contains the number of characters
a symbol applies to in the first column and the relevant symbol in
the second column
}
\description{
This function takes the CIGAR string and converts it into a dataframe
that is able to be read by downstream functions.
}
\seealso{
\code{\link{pipeline}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
