% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{fileCheck}
\alias{fileCheck}
\title{File Checking}
\usage{
fileCheck(filename)
}
\arguments{
\item{filename}{string specifying the path to the user file.}
}
\value{
The resulting path to BAM file
}
\description{
This function ensures that the file specified by the user exists
and is of an accepted type, and gives the file path. If SAM, the
file path returned is for the newly created BAM file.
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
