% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{genomeCheck}
\alias{genomeCheck}
\title{Genome Checking}
\usage{
genomeCheck(gnm)
}
\arguments{
\item{gnm}{string specifying the reference genome.}
}
\value{
The resulting genome object
}
\description{
This function ensures that the reference genome specified by the
user exists and is of an accepted type, and creates the genome
object in R.
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
