% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline}
\usage{
pipeline(read, makedf, gnm)
}
\arguments{
\item{read}{row from BAM file corresponding to the read being aligned}

\item{gnm}{reference genome reads are aligned to}
}
\value{
Nothing
}
\description{
This function gets the relevant information from the read,
generates the CIGAR dataframe, finds the corresponding reference
sequence off the reference genome, accounts for any hard clipping
present at the start of a read, and sends all of this information
the the alignment function.
}
\seealso{
\code{\link{cigarReader}}\cr
\code{\link{alignment}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
