% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlign_PACKAGE.R
\docType{package}
\name{readAlign-package}
\alias{readAlign-package}
\alias{readAlign_PACKAGE}
\alias{readAlign}
\title{readAlign -  shows the precise alignment of a read to the reference
genome according to the mapped read’s CIGAR string.}
\description{
The \code{readAlign} package functions to show the precise alignment of
one or all reads from a BAM/SAM file to the reference genome.
}
\details{
\tabular{ll}{
Package: \tab readAlign\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2024-06-13\cr
License: \tab GPL (>=2)\cr
}
}
\references{
Morgan M, Pagès H, Obenchain V, Hayden N (2024). \emph{Rsamtools:\cr
Binary alignment (BAM), FASTA, variant call (BCF), and tabix\cr
file import}. R package version 2.20.0,\cr
\url{https://bioconductor.org/packages/Rsamtools}\cr
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings:\cr
Efficient manipulation of biological strings}. R package\cr
version 2.72.0, \url{https://bioconductor.org/packages/Biostrings}\cr
Pagès H (2024). \emph{BSgenome: Software infrastructure for\cr
efficient representation of full genomes and their SNPs}. R\cr
package version 1.72.0,\cr
\url{https://bioconductor.org/packages/BSgenome}\cr
Lawrence M, Huber W, Pages H, Aboyoun P, Carlson M, et al. \cr
(2013) Software for Computing and Annotating Genomic Ranges.\cr
PLoS Comput Biol 9(8): e1003118.\cr
doi:10.1371/journal.pcbi.1003118\cr
Lawrence M, Huber W, Pages H, Aboyoun P, Carlson M, et al.\cr
(2013) Software for Computing and Annotating Genomic Ranges.\cr
PLoS Comput Biol 9(8): e1003118.\cr
doi:10.1371/journal.pcbi.1003118
}
\author{
Fateema Bazzi\cr
Politecnico di Milano\cr
Maintainer: Fateema Bazzi\cr
E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
