% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\name{showAlignment}
\alias{showAlignment}
\title{Show the Alignment}
\usage{
showAlignment(cigar)
}
\arguments{
\item{cigar}{CIGAR string associated with a read}
}
\value{
Nothing
}
\description{
In the case of only one read being considered, this function prints
the CIGAR string along with the aligned read to the reference genome.
}
\seealso{
\code{\link{alignment}}\cr
}
\author{
Fateema Bazzi\cr Politecnico di Milano\cr Maintainer: Fateema
Bazzi\cr E-Mail: \href{mailto:fateemahani.bazzi@mail.polimi.it}{fateemahani.bazzi@mail.polimi.it}
}
\keyword{internal}
